Moved results_db.py under /sqlalchemy
